+++ src/functions.ts
@@ -1,35 +1,21 @@
 import * as vscode from 'vscode';
-import {LANGUAGES} from './helpers';
 import type {Languages} from './types';
+import {getFileLanguage, getDecoration, getConfigSettings, LANGUAGES} from './helpers';
 
-// ------------------------------------------------------ //
-// ---------------- BUILD SUBTITLE STRING --------------- //
-// ------------------------------------------------------ //
-
+/**
+ *
+ * @param rawTitle Takes a string and returns a formatted subtitle all UpperCase
+ * @param language Takes file language as string and returns the comment syntax for open and close tags
+ * @returns A formatted subtitle string
+ */
 export const getSubtitle = (rawTitle: string, language: keyof Languages): string => {
-    //
-    // ----------- GETTINGS CONFIG SETTINGS VALUES ---------- //
+    // --------------- GETTING SETTINGS CONFIG -------------- //
 
-    const config = vscode.workspace.getConfiguration('code-titler');
-    const length = config.get<number>('length', 60);
-    const fill = config.get<string>('fill', '-');
-    const personalisedTags = config.get<boolean>('personalisedTags', false);
-    const openTag = config.get<string>('openTag', '//');
-    const closeTag = config.get<string>('closeTag', '//');
-
-    // --------------- GET OPEN AND CLOSE TAGS -------------- //
-
-    let {open, close} = LANGUAGES[language];
-
-    // ------- IF PERSONALISED TAGS CHANGE ACCORDINGLY ------ //
-
-    if (personalisedTags) {
-        open = openTag;
-        close = closeTag;
-    }
+    const {personalisedTags, fill, length} = getConfigSettings();
+    const languageSelector = personalisedTags ? 'personalised' : language;
+    let {open, close} = LANGUAGES[languageSelector];
 
     // ------------------ GET SUBTITLE SIZE ----------------- //
-    // ------------- FOR FORMATTING CALCULATIONS ------------ //
 
     const title = rawTitle.trim();
     const textLength = title.length;
@@ -46,35 +32,71 @@ export const getSubtitle = (rawTitle: string, language: keyof Languages): string
     return subtitleOutput;
 };
 
-// ------------------------------------------------------ //
-// ----------------- BUILD TITLE STRING ----------------- //
-// ------------------------------------------------------ //
-
+/**
+ *
+ * @param rawTitle Takes a string and returns a formatted title
+ * @param language Takes file language as string and returns the comment syntax for open and close tags
+ * @returns A formatted title string
+ */
 export const getTitle = (rawTitle: string, language: keyof Languages): string => {
-    //
-    // ----------- GETTINGS CONFIG SETTINGS VALUES ---------- //
+    // --------------- GETTING SETTINGS CONFIG -------------- //
 
-    const config = vscode.workspace.getConfiguration('code-titler');
-    const length = config.get<number>('length', 60);
-    const fill = config.get<string>('fill', '-');
-    const personalisedTags = config.get<boolean>('personalisedTags', false);
-    const openTag = config.get<string>('openTag', '//');
-    const closeTag = config.get<string>('closeTag', '//');
-
-    // --------------- GET OPEN AND CLOSE TAGS -------------- //
-
-    let {open, close} = LANGUAGES[language];
-
-    // ------- IF PERSONALISED TAGS CHANGE ACCORDINGLY ------ //
-
-    if (personalisedTags) {
-        open = openTag;
-        close = closeTag;
-    }
+    const {personalisedTags, fill, length} = getConfigSettings();
+    const languageSelector = personalisedTags ? 'personalised' : language;
+    let {open, close} = LANGUAGES[languageSelector];
 
     // --------------------- BUILD TITLE -------------------- //
 
     const topBottom = `${open} ${fill.repeat(length - 6)} ${close}`;
     const titleOutput = `${topBottom}\n${getSubtitle(rawTitle, language)}\n${topBottom}`;
     return titleOutput;
+};
+
+/**
+ * Decorates the titles and subtitles in the active text editor
+ * @param editor Takes the active text editor and decorates the titles and subtitles
+ */
+export const decorateTitlesAndSubtitles = (editor: vscode.TextEditor): void => {
+    // --------------- GETTING SETTINGS CONFIG -------------- //
+    const config = vscode.workspace.getConfiguration('code-titler');
+    const personalisedTags = config.get<boolean>('personalisedTags', false);
+    const fill = config.get<string>('fill', '-');
+
+    const documentText = editor.document;
+    const documentLines = documentText.lineCount;
+
+    // --------- GET CURRENT LANGUAGE FILL AND TAGS --------- //
+
+    const language = getFileLanguage(editor);
+
+    const languageSelector = personalisedTags ? 'personalised' : language;
+    let {open, close} = LANGUAGES[languageSelector];
+
+    const openTag = `${open} ${fill}`;
+    const closeTag = `${fill} ${close}`;
+
+    // --------------- CLEAN UP OLD DECORATION -------------- //
+
+    const titleDecoration = getDecoration();
+    const ranges: vscode.Range[] = [];
+
+    // ------------ APPY DECORATION LINE BY LINE ------------ //
+
+    for (let lineIndex = 0; lineIndex < documentLines; lineIndex++) {
+        const currentLine = documentText.lineAt(lineIndex);
+        const lineText = currentLine.text;
+
+        const openTagIndex = lineText.indexOf(openTag);
+        const closeTagIndex = lineText.indexOf(closeTag);
+
+        if (openTagIndex === -1 || closeTagIndex === -1) continue;
+
+        const startPos = new vscode.Position(lineIndex, openTagIndex);
+        const endPos = new vscode.Position(lineIndex, closeTagIndex + 4);
+        const range = new vscode.Range(startPos, endPos);
+
+        ranges.push(range);
+    }
+
+    editor.setDecorations(titleDecoration, ranges);
 };